Modified the code a bit
Cleaned up the variable declarations. Added initializing values to declarations. Resized the std::cout statements to be multiline and to fit GitHub's page. Added the usage of the <limits> library for clearing the stream. Replaced some IF statements with ternary statements, and moved them inside of the std::cout statements themselves. Removed the "return 0;" statement at the end, since it is not required in the main() function. Added a couple of comments (Yuck!!!).